+++ EventListener/ExecListener.php
@@ -1,5 +1,14 @@
 <?php
 
+/*
+ * This file is part of the SpBowerBundle package.
+ *
+ * (c) Martin Parsiegla <martin.parsiegla@gmail.com>
+ *
+ * For the full copyright and license information, please view the LICENSE
+ * file that was distributed with this source code.
+ */
+
 namespace Sp\BowerBundle\EventListener;
 
 use Sp\BowerBundle\Bower\BowerEvent;